+++ utils.h
@@ -1,13 +1,11 @@
 #ifndef ETRC21_UTILS_H_
 #define ETRC21_UTILS_H_
 
-#include "ev3api.h"
-
 class PidControl {
  public:
   PidControl();
   void SetGain(float kp, float ki, float kd);
-  float GetMv(int32_t target_val, int32_t current_val);
+  float GetMv(float target_val, float current_val);
 
  private:
   float kp_;
